Split functionality between the abstract template class Polynomial and the RPolynomial class.
