+++ LinderLab1A-master/LinderLab1A/PolynomialElement.h
@@ -4,8 +4,10 @@
 #include "PositiveNumber.h"
 #include "SignedNumber.h"
 
+// Implemented by M.Tyshchenko and Y. Kishchuk
 template<typename TCoefficient>
 class PolynomialElement {
+    // Implemented by Y. Kishchuk
     static_assert(std::is_member_function_pointer<decltype(&TCoefficient::toString)>::value,
         "Type TCoefficient does not have a toString member function.");
     static_assert(std::is_same<decltype(std::declval<TCoefficient>() + std::declval<TCoefficient>()), TCoefficient>::value,
@@ -16,6 +18,7 @@ class PolynomialElement {
         "Type TCoefficient does not have an operator* defined.");
 
 private:
+    // Implemented by M.Tyshchenko
     PositiveNumber degree;
     TCoefficient coefficient;
 
@@ -60,6 +63,7 @@ public:
         return coef + "x^" + degree.toString();
     }
 
+    // Implemented by Y. Kishchuk
     PolynomialElement operator+(const PolynomialElement& right) const {
         assert(this->degree == right.degree, "Elements have different degree");
         return PolynomialElement(this->coefficient + right.coefficient, this->degree);
